refactor(prompts): group prompt internals
