+++ lib/langfuse.rb
@@ -41,13 +41,13 @@ end
 
 require_relative "langfuse/config"
 require_relative "langfuse/cache_constants"
-require_relative "langfuse/prompt_cache"
-require_relative "langfuse/prompt_fetch_result"
+require_relative "langfuse/prompts/prompt_cache"
+require_relative "langfuse/prompts/prompt_fetch_result"
 require_relative "langfuse/rails_cache_adapter"
-require_relative "langfuse/prompt_cache_capabilities"
-require_relative "langfuse/prompt_cache_coordinator"
+require_relative "langfuse/prompts/prompt_cache_capabilities"
+require_relative "langfuse/prompts/prompt_cache_coordinator"
 require_relative "langfuse/cache_warmer"
-require_relative "langfuse/prompt_cache_events"
+require_relative "langfuse/prompts/prompt_cache_events"
 require_relative "langfuse/api_client"
 require_relative "langfuse/span_filter"
 require_relative "langfuse/sampling"
@@ -59,11 +59,11 @@ require_relative "langfuse/span_processor"
 require_relative "langfuse/observations"
 require_relative "langfuse/trace_id"
 require_relative "langfuse/score_client"
-require_relative "langfuse/prompt_renderer"
-require_relative "langfuse/prompt_client_metadata"
+require_relative "langfuse/prompts/prompt_renderer"
+require_relative "langfuse/prompts/prompt_client_metadata"
 require_relative "langfuse/text_prompt_client"
 require_relative "langfuse/chat_prompt_client"
-require_relative "langfuse/prompt_client_factory"
+require_relative "langfuse/prompts/prompt_client_factory"
 require_relative "langfuse/timestamp_parser"
 require_relative "langfuse/evaluation"
 require_relative "langfuse/experiment_item"
