+++ lib/langfuse/api_client.rb
@@ -5,9 +5,9 @@ require "faraday/retry"
 require "base64"
 require "json"
 require "uri"
-require_relative "prompt_fetch_result"
-require_relative "prompt_cache_capabilities"
-require_relative "prompt_cache_coordinator"
+require_relative "prompts/prompt_fetch_result"
+require_relative "prompts/prompt_cache_capabilities"
+require_relative "prompts/prompt_cache_coordinator"
 require_relative "resources/batches"
 require_relative "resources/datasets"
 require_relative "resources/prompts"
