+++ lib/langfuse/prompts/prompt_cache_capabilities.rb
