+++ lib/langfuse/prompts/prompt_cache_coordinator.rb
