+++ lib/langfuse/prompts/prompt_client_metadata.rb
