+++ lib/langfuse/prompts/prompt_renderer.rb
